chore:implemented paginated-view in views
