+++ wsgi/odyssy/basic/utils.py
@@ -1,5 +1,6 @@
 from datetime import datetime, time
 
+from django.core.paginator import Paginator, EmptyPage
 from django.utils import timezone
 from django.db.models import Q
 
@@ -18,3 +19,14 @@ def get_today_start():
 def get_today_end():
     """ set end as undefined date """
     return datetime.combine(datetime.today(), time.max)
+
+
+def paginate_view(request, query_set, num_items):
+    """ Paginates view from queryset """
+    paginator = Paginator(query_set, num_items)
+    page = request.GET.get('page', default=1)
+    try:
+        data_set = paginator.page(page)
+    except EmptyPage:
+        data_set = paginator.page(paginator.num_pages)
+    return data_set
