+++ wsgi/odyssy/events/views.py
@@ -1,4 +1,5 @@
-from django.core.paginator import Paginator, EmptyPage
+from basic import utils
+
 from django.shortcuts import render
 from .models import Event
 
@@ -6,12 +7,7 @@ from .models import Event
 def event(request):
     events_list = Event.get_all_events()
     num_items = request.GET.get('num_items', default=10)
-    paginator = Paginator(events_list, num_items)
-    page = request.GET.get('page', default=1)
-    try:
-        events = paginator.page(page)
-    except EmptyPage:
-        events = paginator.page(paginator.num_pages)
+    events = utils.paginate_view(request, events_list, num_items)
     context = {'events': events, 'num_items': num_items}
     return render(request, 'events/event_list.html', context=context)
