+++ wsgi/odyssy/news/views.py
@@ -1,4 +1,5 @@
-from django.core.paginator import Paginator, EmptyPage
+from basic import utils
+
 from django.shortcuts import render
 from .models import News
 
@@ -6,12 +7,7 @@ from .models import News
 def news(request):
     news_list = News.get_all_news()
     num_items = request.GET.get('num_items', default=10)
-    paginator = Paginator(news_list, num_items)
-    page = request.GET.get('page', default=1)
-    try:
-        news_context = paginator.page(page)
-    except EmptyPage:
-        news_context = paginator.page(paginator.num_pages)
+    news_context = utils.paginate_view(request, news_list, num_items)
     context = {
         'news_list': news_context,
         'num_items': num_items
