+++ wsgi/odyssy/tag/views.py
@@ -1,9 +1,9 @@
 from announcement.models import Announcement
 from news.models import News
 from basic.models import PhotoExtended
+from basic import utils
 
 from django.shortcuts import render
-from django.core.paginator import Paginator, EmptyPage
 
 
 def simple_tag(request, tag_name):
@@ -19,21 +19,10 @@ def simple_tag(request, tag_name):
     return render(request, 'tag/simple_tag.html', args)
 
 
-def paginate_view(request, query_set, num_items):
-    """ Paginates view from queryset """
-    paginator = Paginator(query_set, num_items)
-    page = request.GET.get('page', default=1)
-    try:
-        data_set = paginator.page(page)
-    except EmptyPage:
-        data_set = paginator.page(paginator.num_pages)
-    return data_set
-
-
 def announcement_tag(request, tag_name):
     """ View for announcement tag listing """
     announcements = Announcement.get_announcement_tag(tag_name)
-    announcements = paginate_view(request, announcements, 10)
+    announcements = utils.paginate_view(request, announcements, 10)
     args = {'announcements': announcements}
     return render(request, 'tag/announcement_tag.html', args)
 
@@ -41,7 +30,7 @@ def announcement_tag(request, tag_name):
 def news_tag(request, tag_name):
     """ View for news tag listing """
     news = News.get_news_tag(tag_name)
-    news = paginate_view(request, news, 10)
+    news = utils.paginate_view(request, news, 10)
     args = {'news': news}
     return render(request, 'tag/news_tag.html', args)
 
@@ -49,6 +38,6 @@ def news_tag(request, tag_name):
 def picture_tag(request, tag_name):
     """ View for picture tag listing """
     images = PhotoExtended.get_photo_tag(tag_name)
-    images = paginate_view(request, images, 10)
+    images = utils.paginate_view(request, images, 10)
     args = {'images': images}
     return render(request, 'tag/picture_tag.html', args)
